Add some auto defaults + updated electron version
This should work but I haven't tested it.
Added a dynamic Electron Version by seeing if its in dependencies, dev dependencies and falling back to v7.1.5.
Added a dynamic version check by using the version in the package.json falling back to v0.0.1.
Added a default icon file, which is where I store my icon as well as I've seen someone else do so. This change is more for my convenience but it is a smart place to save your icon.